+++ lib/ts/recipe/thirdpartyemailpassword/types.ts
@@ -12,7 +12,7 @@
  * License for the specific language governing permissions and limitations
  * under the License.
  */
-import { TypeProvider, TypeProviderGetResponse } from "../thirdparty/types";
+import { TypeProvider } from "../thirdparty/types";
 import { TypeInput as TypeNormalisedInputEmailVerification } from "../emailverification/types";
 import {
     NormalisedFormField,